Broke out wasm module decoder.rs into seperate .rs files
